+++ resources/views/livewire/pages/role/index.blade.php
@@ -1,10 +1,12 @@
 <div class="space-y-6">
     <div class="flex justify-between">
         <input type="search" class="input input-bordered" placeholder="Cari permission" wire:model.live="cari">
-        <button class="btn btn-primary" wire:click="$dispatch('addPermission')">
-            <x-tabler-circle-plus class="icon-5" />
-            <span>Create Role or permission</span>
-        </button>
+        @can('role.create')
+            <button class="btn btn-primary" wire:click="$dispatch('addPermission')">
+                <x-tabler-circle-plus class="icon-5" />
+                <span>Create Role or permission</span>
+            </button>
+        @endcan
     </div>
     <div class="table-wrapper">
         <table class="table">
@@ -14,7 +16,9 @@
                 @foreach ($roles as $role)
                     <th class="text-center capitalize">{{ $role->name }}</th>
                 @endforeach
-                <th class="text-center">Action</th>
+                @canany(['role.edit', 'role.delete'])
+                    <th class="text-center">Action</th>
+                @endcanany
             </thead>
             <tbody>
                 @forelse ($permissions as $permit)
@@ -23,23 +27,39 @@
                         <td>{{ $permit->name }}</td>
                         @foreach ($roles as $role)
                             <td class="text-center">
-                                <input type="checkbox" class="toggle toggle-primary toggle-sm"
-                                    @checked($permit->hasRole($role->name))
-                                    wire:change="applyPermission({{ $permit->id }}, '{{ $role->name }}')" />
+                                <div class="flex items-center justify-center">
+                                    @can('role.edit')
+                                        <input type="checkbox" class="toggle toggle-primary toggle-sm"
+                                            @checked($permit->hasRole($role->name))
+                                            wire:change="applyPermission({{ $permit->id }}, '{{ $role->name }}')" />
+                                    @else
+                                        <x-tabler-check @class([
+                                            'icon-5',
+                                            'stroke-primary block' => $permit->hasRole($role->name),
+                                            'hidden' => !$permit->hasRole($role->name),
+                                        ]) />
+                                    @endcan
+                                </div>
                             </td>
                         @endforeach
-                        <td>
-                            <div class="flex gap-1 justify-center">
-                                <button class="btn input-bordered btn-xs btn-square"
-                                    wire:click="$dispatch('editPermission', {permission: {{ $permit->id }}})">
-                                    <x-tabler-edit class="icon-4" />
-                                </button>
-                                <button class="btn input-bordered btn-xs btn-square"
-                                    wire:click="deletePermission({{ $permit->id }})">
-                                    <x-tabler-trash class="icon-4" />
-                                </button>
-                            </div>
-                        </td>
+                        @canany(['role.edit', 'role.delete'])
+                            <td>
+                                <div class="flex gap-1 justify-center">
+                                    @can('role.edit')
+                                        <button class="btn input-bordered btn-xs btn-square"
+                                            wire:click="$dispatch('editPermission', {permission: {{ $permit->id }}})">
+                                            <x-tabler-edit class="icon-4" />
+                                        </button>
+                                    @endcan
+                                    @can('role.delete')
+                                        <button class="btn input-bordered btn-xs btn-square"
+                                            wire:click="deletePermission({{ $permit->id }})">
+                                            <x-tabler-trash class="icon-4" />
+                                        </button>
+                                    @endcan
+                                </div>
+                            </td>
+                        @endcanany
                     </tr>
                 @empty
                     <tr>
